Update orderGroup mock queries
