+++ react/__tests__/Header.test.js
@@ -3,13 +3,13 @@ import React from 'react'
 import { render } from '../testUtils'
 import Header from '../components/Header'
 import Warnings from '../components/Header/Warnings'
-import { orderGroupQuery as oneDelivery } from '../mocks/1-delivery'
-import { orderGroupQuery as onePickup } from '../mocks/1-pickup'
-import { orderGroupQuery as splitOrder } from '../mocks/split-order-2-sellers'
-import { orderGroupQuery as bankInvoiceDueDate } from '../mocks/mockBoletoLoggedIn'
-import { orderGroupQuery as bankInvoiceNumber } from '../mocks/mockBoletoNumberLoggedIn'
-import { orderGroupQuery as bankInvoiceNoDueDate } from '../mocks/bankinvoice'
-import { orderGroupQuery as deliveryAndPickup } from '../mocks/3-pickup-only-2-delivery-only'
+import { orderGroupQuery as oneDelivery } from '../mocks/oneDeliverySimple'
+import { orderGroupQuery as onePickup } from '../mocks/onePickupSimple'
+import { orderGroupQuery as splitOrder } from '../mocks/splitOrderTwoSellers'
+import { orderGroupQuery as bankInvoiceDueDate } from '../mocks/bankInvoiceLoggedIn'
+import { orderGroupQuery as bankInvoiceNumber } from '../mocks/bankInvoiceNumberLoggedIn'
+import { orderGroupQuery as bankInvoiceNoDueDate } from '../mocks/bankInvoice'
+import { orderGroupQuery as deliveryAndPickup } from '../mocks/pickupAndDelivery'
 
 describe('Confirmation messages', () => {
   it('should render success icon', () => {
