--- react/mocks/additionalInfoProducts.js
@@ -1,623 +1,0 @@
-export const OrderGroupQuery = {
-  orderGroup: [
-    {
-      allowCancellation: true,
-      orderId: '904272889752-01',
-      orderGroup: '904272889752',
-      state: 'payment-approved',
-      salesChannel: '1',
-      creationDate: '2019-01-18T18:47:54.66935Z',
-      timeZoneCreationDate: '2019-01-18T16:47:54.66935',
-      value: 27903,
-      clientProfileData: {
-        email: 'minidog@mailinator.com',
-        firstName: 'Victor',
-        lastName: 'Hugo',
-        document: '12345678909',
-        documentType: 'cpf',
-        phone: '+551199999999',
-      },
-      storePreferencesData: {
-        currencyCode: 'BRL',
-        countryCode: 'BRA',
-      },
-      sellers: [
-        {
-          id: '1',
-          name: 'vtexgame1',
-        },
-      ],
-      totals: [
-        {
-          id: 'Items',
-          name: 'Total dos Itens',
-          value: 26063,
-        },
-        {
-          id: 'Discounts',
-          name: 'Total dos Descontos',
-          value: 0,
-        },
-        {
-          id: 'Shipping',
-          name: 'Total do Frete',
-          value: 1840,
-        },
-        {
-          id: 'Tax',
-          name: 'Total da Taxa',
-          value: 0,
-        },
-      ],
-      items: [
-        {
-          uniqueId: 'B7E21B3C1E0D44E98888AE01383C3B68',
-          id: '34',
-          productId: '23',
-          name: 'Camisa Vasco Branca',
-          skuName: 'Camisa Vasco Branca',
-          tax: 0,
-          price: 920,
-          listPrice: 1000,
-          sellingPrice: 920,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155633-55-55/camisavascobranca.jpg?v=636517252260000000',
-          detailUrl: '/camisavascobranca/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1,
-        },
-        {
-          uniqueId: 'B7E21B3C1E0D44E98888AE01383C3B68',
-          id: '34',
-          productId: '23',
-          name: 'Camisa Vasco Branca',
-          skuName: 'Camisa Vasco Branca',
-          tax: 0,
-          price: 920,
-          listPrice: 1000,
-          sellingPrice: 920,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155633-55-55/camisavascobranca.jpg?v=636517252260000000',
-          detailUrl: '/camisavascobranca/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1,
-        },
-        {
-          uniqueId: '45497AF8348847EFB7AA0F44BBB7FBF1',
-          id: '260',
-          productId: '244',
-          name: 'Camisa Sao Paulo',
-          skuName: 'Camisa Sao Paulo',
-          tax: 0,
-          price: 1000,
-          listPrice: 1000,
-          sellingPrice: 1000,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155638-55-55/camisasaopaulo.jpg?v=636519608956130000',
-          detailUrl: '/camisa-sao-paulo/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1,
-        },
-        {
-          uniqueId: '5D114BC03D774A2494BDD390F30F3BBC',
-          id: '261',
-          productId: '245',
-          name: 'Camisa Grêmio',
-          skuName: 'Camisa Grêmio',
-          tax: 0,
-          price: 19990,
-          listPrice: 19990,
-          sellingPrice: 19990,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155628-55-55/gremio1.jpg?v=636480000752030000',
-          detailUrl: '/camisagremio/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1,
-        },
-        {
-          uniqueId: 'D49E12C4600C4DD9B61634763638BF46',
-          id: '253',
-          productId: '238',
-          name: 'Principal e franquia franquia22',
-          skuName: 'franquia22',
-          tax: 0,
-          price: 500,
-          listPrice: 500,
-          sellingPrice: 500,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155609-55-55/sku_principal_franquia.png?v=636433217889570000',
-          detailUrl: '/principal-franquia/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1,
-        },
-        {
-          uniqueId: 'B9EFF098448242E189C65277EB8F3415',
-          id: '40',
-          productId: '29',
-          name:
-            'Produto com desconto percentual por forma de pagamento Sku com desconto percentual por forma de pagamento boleto',
-          skuName: 'Sku com desconto percentual por forma de pagamento boleto',
-          tax: 0,
-          price: 1200,
-          listPrice: 1500,
-          sellingPrice: 1200,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155377-55-55/Produto-Desconto-percentual-por-forma-de-pagamento-boleto.png?v=635930330014300000',
-          detailUrl:
-            '/produto-com-desconto-percentual-por-forma-de-pagamento-boleto/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1,
-        },
-        {
-          uniqueId: '99D6812E06FF4707A2EBB0F6002CE82E',
-          id: '17',
-          productId: '6',
-          name: 'Produto unidade de medida Sku unidade de medida',
-          skuName: 'Sku unidade de medida',
-          tax: 0,
-          price: 400,
-          listPrice: 530,
-          sellingPrice: 533,
-          isGift: false,
-          quantity: 1,
-          seller: '1',
-          imageUrl:
-            'http://vtexgame1.vteximg.com.br/arquivos/ids/155352-55-55/Produto-undidade-de-medida.png?v=635812139082600000',
-          detailUrl: '/produtounidadedemedida/p',
-          measurementUnit: 'un',
-          unitMultiplier: 1.3337,
-        },
-      ],
-      shippingData: {
-        selectedAddresses: [
-          {
-            addressId: '61910219929760',
-            addressType: 'residential',
-            receiverName: 'Victor Hugo',
-            city: 'Rio de Janeiro',
-            state: 'RJ',
-            street: 'Praia de Botafogo',
-            number: '3000',
-            neighborhood: 'Botafogo',
-            complement: null,
-            postalCode: '22250-040',
-            country: 'BRA',
-          },
-        ],
-        logisticsInfo: [
-          {
-            itemIndex: 0,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'agendada',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'agendada',
-                name: 'agendada',
-                price: 170,
-                shippingEstimate: '0bd',
-                deliveryWindow: {
-                  startDateUtc: '2019-01-22T09:00:00+00:00',
-                  endDateUtc: '2019-01-22T21:00:00+00:00',
-                },
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-          {
-            itemIndex: 1,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'agendada',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'agendada',
-                name: 'agendada',
-                price: 170,
-                shippingEstimate: '0bd',
-                deliveryWindow: {
-                  startDateUtc: '2019-01-22T09:00:00+00:00',
-                  endDateUtc: '2019-01-22T21:00:00+00:00',
-                },
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-          {
-            itemIndex: 2,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'Econômica',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'Econômica',
-                name: 'Econômica',
-                price: 0,
-                shippingEstimate: '10bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'Normal',
-                name: 'Normal',
-                price: 100,
-                shippingEstimate: '8bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'PAC',
-                name: 'PAC',
-                price: 250,
-                shippingEstimate: '3bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-          {
-            itemIndex: 3,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'Econômica',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'Econômica',
-                name: 'Econômica',
-                price: 0,
-                shippingEstimate: '10bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'Normal',
-                name: 'Normal',
-                price: 100,
-                shippingEstimate: '8bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'PAC',
-                name: 'PAC',
-                price: 250,
-                shippingEstimate: '3bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'Motoboy',
-                name: 'Motoboy',
-                price: 1000,
-                shippingEstimate: '60m',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-          {
-            itemIndex: 4,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'PAC',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'PAC',
-                name: 'PAC',
-                price: 26,
-                shippingEstimate: '13bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'Expressa',
-                name: 'Expressa',
-                price: 52,
-                shippingEstimate: '12bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'retirada na loja (141125d)',
-                name: 'retirada na loja (141125d)',
-                price: 0,
-                shippingEstimate: '10bd',
-                deliveryWindow: null,
-                deliveryChannel: 'pickup-in-point',
-                pickupStoreInfo: {
-                  additionalInfo: 'traga a vasilha\n\nnão se esqueça',
-                  address: {
-                    addressId: '141125d',
-                    addressType: 'pickup',
-                    receiverName: null,
-                    city: 'Rio de Janeiro',
-                    state: 'RJ',
-                    street: 'Rua General Azevedo Pimentel',
-                    number: '5',
-                    neighborhood: 'Copacabana',
-                    complement: '',
-                    postalCode: '22011050',
-                    country: 'BRA',
-                  },
-                  dockId: '12db23b',
-                  friendlyName: 'Loja em Copacabana no Rio de Janeiro',
-                  isPickupStore: true,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-          {
-            itemIndex: 5,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'PAC',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'PAC',
-                name: 'PAC',
-                price: 88,
-                shippingEstimate: '13bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'Expressa',
-                name: 'Expressa',
-                price: 175,
-                shippingEstimate: '12bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'retirada na loja (141125d)',
-                name: 'retirada na loja (141125d)',
-                price: 0,
-                shippingEstimate: '10bd',
-                deliveryWindow: null,
-                deliveryChannel: 'pickup-in-point',
-                pickupStoreInfo: {
-                  additionalInfo: 'traga a vasilha\n\nnão se esqueça',
-                  address: {
-                    addressId: '141125d',
-                    addressType: 'pickup',
-                    receiverName: null,
-                    city: 'Rio de Janeiro',
-                    state: 'RJ',
-                    street: 'Rua General Azevedo Pimentel',
-                    number: '5',
-                    neighborhood: 'Copacabana',
-                    complement: '',
-                    postalCode: '22011050',
-                    country: 'BRA',
-                  },
-                  dockId: '12db23b',
-                  friendlyName: 'Loja em Copacabana no Rio de Janeiro',
-                  isPickupStore: true,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-          {
-            itemIndex: 6,
-            shippingEstimate: null,
-            shippingEstimateDate: null,
-            selectedSla: 'PAC',
-            deliveryChannel: null,
-            addressId: '61910219929760',
-            slas: [
-              {
-                id: 'PAC',
-                name: 'PAC',
-                price: 386,
-                shippingEstimate: '13bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'Expressa',
-                name: 'Expressa',
-                price: 773,
-                shippingEstimate: '12bd',
-                deliveryWindow: null,
-                deliveryChannel: 'delivery',
-                pickupStoreInfo: {
-                  additionalInfo: null,
-                  address: null,
-                  dockId: null,
-                  friendlyName: null,
-                  isPickupStore: false,
-                },
-              },
-              {
-                id: 'retirada na loja (141125d)',
-                name: 'retirada na loja (141125d)',
-                price: 0,
-                shippingEstimate: '10bd',
-                deliveryWindow: null,
-                deliveryChannel: 'pickup-in-point',
-                pickupStoreInfo: {
-                  additionalInfo: 'traga a vasilha\n\nnão se esqueça',
-                  address: {
-                    addressId: '141125d',
-                    addressType: 'pickup',
-                    receiverName: null,
-                    city: 'Rio de Janeiro',
-                    state: 'RJ',
-                    street: 'Rua General Azevedo Pimentel',
-                    number: '5',
-                    neighborhood: 'Copacabana',
-                    complement: '',
-                    postalCode: '22011050',
-                    country: 'BRA',
-                  },
-                  dockId: '12db23b',
-                  friendlyName: 'Loja em Copacabana no Rio de Janeiro',
-                  isPickupStore: true,
-                },
-              },
-            ],
-            deliveryIds: null,
-          },
-        ],
-      },
-      packageAttachment: null,
-      paymentData: {
-        transactions: [
-          {
-            transactionId: '83DAE826352C4C9BB0AAC6A638E04614',
-            payments: [
-              {
-                paymentSystem: '6',
-                paymentSystemName: 'Boleto Bancário',
-                value: 27903,
-                lastDigits: null,
-                group: 'bankInvoice',
-                installments: 1,
-                dueDate: null,
-                url: null,
-                tid: '32885634',
-                bankIssuedInvoiceIdentificationNumberFormatted: null,
-                bankIssuedInvoiceBarCodeNumber: null,
-                bankIssuedInvoiceBarCodePNG: null,
-              },
-            ],
-          },
-        ],
-      },
-    },
-  ],
-}
