+++ react/mocks/onePickupSimple.js
@@ -1,0 +1,184 @@
+export const orderGroupQuery = {
+  orderGroup: [
+    {
+      allowCancellation: true,
+      orderId: '905691315299-01',
+      orderGroup: '905691315299',
+      state: 'payment-pending',
+      salesChannel: '1',
+      creationDate: '2019-01-24T17:21:30.9124556Z',
+      timeZoneCreationDate: '2019-01-24T15:21:30.9124556',
+      value: 30000,
+      clientProfileData: {
+        email: 'victorhmp@mailinator.com',
+        firstName: 'Victor',
+        lastName: 'Hugo',
+        document: '18430995005',
+        documentType: 'cpf',
+        phone: '+551112340909',
+      },
+      storePreferencesData: {
+        currencyCode: 'BRL',
+        countryCode: 'BRA',
+      },
+      sellers: [
+        {
+          id: '1',
+          name: 'vtexgame1',
+        },
+      ],
+      totals: [
+        {
+          id: 'Items',
+          name: 'Total dos Itens',
+          value: 30000,
+        },
+        {
+          id: 'Discounts',
+          name: 'Total dos Descontos',
+          value: 0,
+        },
+        {
+          id: 'Shipping',
+          name: 'Total do Frete',
+          value: 0,
+        },
+        {
+          id: 'Tax',
+          name: 'Total da Taxa',
+          value: 0,
+        },
+      ],
+      items: [
+        {
+          uniqueId: '4ABAD28312824A54A0A405531267CCA0',
+          id: '307',
+          productId: '276',
+          name: 'Pickup múltiplas SLAs RJ Tipo 1',
+          skuName: 'Tipo 1',
+          tax: 0,
+          price: 30000,
+          listPrice: 50000,
+          sellingPrice: 30000,
+          isGift: false,
+          quantity: 1,
+          seller: '1',
+          imageUrl:
+            'http://vtexgame1.vteximg.com.br/arquivos/ids/155657-55-55/prancha.jpg?v=636657805466200000',
+          detailUrl:
+            '/teste-so-delivery-copy-256--copy-257--copy-258--copy-275--copy-276-/p',
+          measurementUnit: 'un',
+          unitMultiplier: 1,
+        },
+      ],
+      shippingData: {
+        selectedAddresses: [
+          {
+            addressId: 'vtexgame1:1f848d7',
+            addressType: 'pickup',
+            receiverName: 'Victor Hugo',
+            city: 'Rio de Janeiro',
+            state: 'RJ',
+            street: 'Rua Marquês de Abrantes',
+            number: '5',
+            neighborhood: 'Flamengo',
+            complement: '',
+            postalCode: '22230061',
+            country: 'BRA',
+          },
+        ],
+        logisticsInfo: [
+          {
+            itemIndex: 0,
+            shippingEstimate: null,
+            shippingEstimateDate: null,
+            selectedSla: 'retirada na loja múltiplos pontos RJ (1f848d7)',
+            deliveryChannel: null,
+            addressId: 'vtexgame1:1f848d7',
+            slas: [
+              {
+                id: 'retirada na loja múltiplos pontos RJ (1f848d7)',
+                name: 'retirada na loja múltiplos pontos RJ (1f848d7)',
+                price: 0,
+                shippingEstimate: '11bd',
+                deliveryWindow: null,
+                deliveryChannel: 'pickup-in-point',
+                pickupStoreInfo: {
+                  additionalInfo: 'não entre de sunga\nnão entre sem sunga',
+                  address: {
+                    addressId: 'vtexgame1:1f848d7',
+                    addressType: 'pickup',
+                    receiverName: null,
+                    city: 'Rio de Janeiro',
+                    state: 'RJ',
+                    street: 'Rua Marquês de Abrantes',
+                    number: '5',
+                    neighborhood: 'Flamengo',
+                    complement: '',
+                    postalCode: '22230061',
+                    country: 'BRA',
+                  },
+                  dockId: '11dbf12',
+                  friendlyName: 'Loja no Flamengo do Rio de Janeiro',
+                  isPickupStore: true,
+                },
+              },
+              {
+                id: 'retirada na loja múltiplos pontos RJ (141125d)',
+                name: 'retirada na loja múltiplos pontos RJ (141125d)',
+                price: 0,
+                shippingEstimate: '11bd',
+                deliveryWindow: null,
+                deliveryChannel: 'pickup-in-point',
+                pickupStoreInfo: {
+                  additionalInfo: 'traga a vasilha\n\nnão se esqueça',
+                  address: {
+                    addressId: '141125d',
+                    addressType: 'pickup',
+                    receiverName: null,
+                    city: 'Rio de Janeiro',
+                    state: 'RJ',
+                    street: 'Rua General Azevedo Pimentel',
+                    number: '5',
+                    neighborhood: 'Copacabana',
+                    complement: '',
+                    postalCode: '22011050',
+                    country: 'BRA',
+                  },
+                  dockId: '11dbf12',
+                  friendlyName: 'Loja em Copacabana no Rio de Janeiro',
+                  isPickupStore: true,
+                },
+              },
+            ],
+            deliveryIds: null,
+          },
+        ],
+      },
+      packageAttachment: null,
+      paymentData: {
+        transactions: [
+          {
+            transactionId: '7D599682F7274DCC9FD34B8386F12123',
+            payments: [
+              {
+                paymentSystem: '4',
+                paymentSystemName: 'Mastercard',
+                value: 30000,
+                lastDigits: '1234',
+                group: 'creditCard',
+                installments: 1,
+                dueDate: null,
+                url: null,
+                tid: '49622574',
+                bankIssuedInvoiceIdentificationNumberFormatted: null,
+                bankIssuedInvoiceBarCodeNumber: null,
+                bankIssuedInvoiceBarCodePNG: null,
+              },
+            ],
+          },
+        ],
+      },
+    },
+  ],
+}
