+++ react/mocks/promissoryPayment.js
@@ -1,0 +1,271 @@
+export const orderGroupQuery = {
+  orderGroup: [
+    {
+      allowCancellation: true,
+      orderId: '905681044769-01',
+      orderGroup: '905681044769',
+      state: 'payment-pending',
+      salesChannel: '1',
+      creationDate: '2019-01-24T16:16:59.8112706Z',
+      timeZoneCreationDate: '2019-01-24T14:16:59.8112706',
+      value: 900,
+      clientProfileData: {
+        email: 'minidog@mailinator.com',
+        firstName: 'Victor',
+        lastName: 'Hugo',
+        document: '12345678909',
+        documentType: 'cpf',
+        phone: '+551199999999',
+      },
+      storePreferencesData: {
+        currencyCode: 'BRL',
+        countryCode: 'BRA',
+      },
+      sellers: [
+        {
+          id: '1',
+          name: 'vtexgame1',
+        },
+      ],
+      totals: [
+        {
+          id: 'Items',
+          name: 'Total dos Itens',
+          value: 900,
+        },
+        {
+          id: 'Discounts',
+          name: 'Total dos Descontos',
+          value: 0,
+        },
+        {
+          id: 'Shipping',
+          name: 'Total do Frete',
+          value: 0,
+        },
+        {
+          id: 'Tax',
+          name: 'Total da Taxa',
+          value: 0,
+        },
+      ],
+      items: [
+        {
+          uniqueId: 'B11E734190A447A7A0DFCAB0F0832464',
+          id: '298',
+          productId: '271',
+          name: 'Pickup 1 SLA e Delivery com várias SLAs Tipo 1',
+          skuName: 'Tipo 1',
+          tax: 0,
+          price: 900,
+          listPrice: 1000,
+          sellingPrice: 900,
+          isGift: false,
+          quantity: 1,
+          seller: '1',
+          imageUrl:
+            'http://vtexgame1.vteximg.com.br/arquivos/ids/155647-55-55/download--1-.jpg?v=636626821858700000',
+          detailUrl:
+            '/teste-so-delivery-copy-256--copy-257--copy-261--copy-262--copy-263--copy-264--copy-265--copy-266--copy-267--copy-268--copy-269--copy-270--copy-271-/p',
+          measurementUnit: 'un',
+          unitMultiplier: 1,
+        },
+      ],
+      shippingData: {
+        selectedAddresses: [
+          {
+            addressId: 'vtexgame1:141125d',
+            addressType: 'pickup',
+            receiverName: 'Victor Hugo',
+            city: 'Rio de Janeiro',
+            state: 'RJ',
+            street: 'Rua General Azevedo Pimentel',
+            number: '5',
+            neighborhood: 'Copacabana',
+            complement: '',
+            postalCode: '22011050',
+            country: 'BRA',
+          },
+        ],
+        logisticsInfo: [
+          {
+            itemIndex: 0,
+            shippingEstimate: null,
+            shippingEstimateDate: null,
+            selectedSla: 'retirada na loja (141125d)',
+            deliveryChannel: null,
+            addressId: 'vtexgame1:141125d',
+            slas: [
+              {
+                id: 'PAC',
+                name: 'PAC',
+                price: 500,
+                shippingEstimate: '10bd',
+                deliveryWindow: null,
+                deliveryChannel: 'delivery',
+                pickupStoreInfo: {
+                  additionalInfo: null,
+                  address: null,
+                  dockId: null,
+                  friendlyName: null,
+                  isPickupStore: false,
+                },
+              },
+              {
+                id: 'Motoboy',
+                name: 'Motoboy',
+                price: 1000,
+                shippingEstimate: '7bd',
+                deliveryWindow: null,
+                deliveryChannel: 'delivery',
+                pickupStoreInfo: {
+                  additionalInfo: null,
+                  address: null,
+                  dockId: null,
+                  friendlyName: null,
+                  isPickupStore: false,
+                },
+              },
+              {
+                id: 'Expressa',
+                name: 'Expressa',
+                price: 1000,
+                shippingEstimate: '9bd',
+                deliveryWindow: null,
+                deliveryChannel: 'delivery',
+                pickupStoreInfo: {
+                  additionalInfo: null,
+                  address: null,
+                  dockId: null,
+                  friendlyName: null,
+                  isPickupStore: false,
+                },
+              },
+              {
+                id: 'PAC Lento',
+                name: 'PAC Lento',
+                price: 1000,
+                shippingEstimate: '22bd',
+                deliveryWindow: null,
+                deliveryChannel: 'delivery',
+                pickupStoreInfo: {
+                  additionalInfo: null,
+                  address: null,
+                  dockId: null,
+                  friendlyName: null,
+                  isPickupStore: false,
+                },
+              },
+              {
+                id: 'retirada na loja múltiplos pontos RJ (1f848d7)',
+                name: 'retirada na loja múltiplos pontos RJ (1f848d7)',
+                price: 0,
+                shippingEstimate: '11bd',
+                deliveryWindow: null,
+                deliveryChannel: 'pickup-in-point',
+                pickupStoreInfo: {
+                  additionalInfo: 'não entre de sunga\nnão entre sem sunga',
+                  address: {
+                    addressId: '1f848d7',
+                    addressType: 'pickup',
+                    receiverName: null,
+                    city: 'Rio de Janeiro',
+                    state: 'RJ',
+                    street: 'Rua Marquês de Abrantes',
+                    number: '5',
+                    neighborhood: 'Flamengo',
+                    complement: '',
+                    postalCode: '22230061',
+                    country: 'BRA',
+                  },
+                  dockId: '11dbf12',
+                  friendlyName: 'Loja no Flamengo do Rio de Janeiro',
+                  isPickupStore: true,
+                },
+              },
+              {
+                id: 'retirada na loja (141125d)',
+                name: 'retirada na loja (141125d)',
+                price: 0,
+                shippingEstimate: '2bd',
+                deliveryWindow: null,
+                deliveryChannel: 'pickup-in-point',
+                pickupStoreInfo: {
+                  additionalInfo: 'traga a vasilha\n\nnão se esqueça',
+                  address: {
+                    addressId: 'vtexgame1:141125d',
+                    addressType: 'pickup',
+                    receiverName: null,
+                    city: 'Rio de Janeiro',
+                    state: 'RJ',
+                    street: 'Rua General Azevedo Pimentel',
+                    number: '5',
+                    neighborhood: 'Copacabana',
+                    complement: '',
+                    postalCode: '22011050',
+                    country: 'BRA',
+                  },
+                  dockId: '1c38481',
+                  friendlyName: 'Loja em Copacabana no Rio de Janeiro',
+                  isPickupStore: true,
+                },
+              },
+              {
+                id: 'retirada na loja múltiplos pontos RJ (141125d)',
+                name: 'retirada na loja múltiplos pontos RJ (141125d)',
+                price: 0,
+                shippingEstimate: '11bd',
+                deliveryWindow: null,
+                deliveryChannel: 'pickup-in-point',
+                pickupStoreInfo: {
+                  additionalInfo: 'traga a vasilha\n\nnão se esqueça',
+                  address: {
+                    addressId: 'vtexgame1:141125d',
+                    addressType: 'pickup',
+                    receiverName: null,
+                    city: 'Rio de Janeiro',
+                    state: 'RJ',
+                    street: 'Rua General Azevedo Pimentel',
+                    number: '5',
+                    neighborhood: 'Copacabana',
+                    complement: '',
+                    postalCode: '22011050',
+                    country: 'BRA',
+                  },
+                  dockId: '1c38481',
+                  friendlyName: 'Loja em Copacabana no Rio de Janeiro',
+                  isPickupStore: true,
+                },
+              },
+            ],
+            deliveryIds: null,
+          },
+        ],
+      },
+      packageAttachment: null,
+      paymentData: {
+        transactions: [
+          {
+            transactionId: 'ADE7DECB757A4DEEB1CD18F27DE93843',
+            payments: [
+              {
+                paymentSystem: '17',
+                paymentSystemName: 'Cash',
+                value: 900,
+                lastDigits: null,
+                group: 'promissory',
+                installments: 1,
+                dueDate: null,
+                url: null,
+                tid: null,
+                bankIssuedInvoiceIdentificationNumberFormatted: null,
+                bankIssuedInvoiceBarCodeNumber: null,
+                bankIssuedInvoiceBarCodePNG: null,
+              },
+            ],
+          },
+        ],
+      },
+    },
+  ],
+}
